GetInstalledApp.ps1
This PowerShell script is designed to extract detailed information about the software installed on a remote computer. It operates by accepting the name of the remote computer as a parameter. The script then accesses the Windows Registry of the specified computer, specifically targeting the registry key that holds data about installed programs. It retrieves all subkeys under this registry key, with each subkey representing an installed program. For each program, the script creates a new object and populates it with properties such as the program’s display name, version, installation location, and publisher. The resulting array of objects, which collectively provide a comprehensive overview of the software environment on the remote computer, is then outputted in a formatted table for easy viewing.